move non-TEE-specific UI components into their own dir
