+++ src/components/generic/SyncIndicator.tsx
